refactor(knowledge-store): extract the converger's row primitives
Row upsert, rename following, resolution and pruning move to index/rows.py
unchanged, so a second writer can share one notion of document identity.
Nothing in there imports the indexing pipeline: the commit path is about to pay
this module's import cost on every save.
